Continuing Anki 2 port changes:  - Anki2 library updates  - unit test fixes  - Qt forms regenerated
Config now stored locally not in an Anki profile as they are not available when the addon is initialised (so seperate config for seperate profiles currently not supported).

NOTE: git master stream is now continued on from the last release Anki1. Tagged at 0.6.3. The battersea master was in a messy state with uncompleted enhancements and bugs. I will check in the work I did trying to get this working into another branch.